# [WIP] Vm inheritance change for type column