--- a/spec/models/vm_spec.rb
+++ b/spec/models/vm_spec.rb
@@ -13,6 +13,12 @@ describe Vm do
     expect(ManageIQ::Providers::Redhat::InfraManager::Vm.corresponding_template_model).to eq(ManageIQ::Providers::Redhat::InfraManager::Template)
   end
 
+  context "#template" do
+    it "defaults to true" do
+      expect(described_class.new.template).to eq(false)
+    end
+  end
+
   context "#template=" do
     before(:each) { @vm = FactoryGirl.create(:vm_vmware) }