@chrisarcand 2017-12-06
I'm confused

> Curiously, we should only need to look at the template value rather than building the long type IN() queries.

Really? Why?

---

I'm pretty wary of this (even if the template reasoning above works out?) as the `IN` calls are how Rails *does* STI, and you're overriding a private API to change this likely very important behavior. Example, to remove us from ManageIQ extensionland:

```ruby
# frozen_string_literal: true

begin
  require "bundler/inline"
rescue LoadError => e
  $stderr.puts "Bundler version 1.10 or later is required. Please update your Bundler"
  raise e
end

gemfile(true) do
  source "https://rubygems.org"

  git_source(:github) { |repo| "https://github.com/#{repo}.git" }

  # Activate the gem you are reporting the issue against.
  gem "activerecord", "5.1.0"
  gem "sqlite3"
  gem 'pry-rails'
end

require "active_record"
require "minitest/autorun"
require "logger"

# Ensure backward compatibility with Minitest 4
Minitest::Test = MiniTest::Unit::TestCase unless defined?(Minitest::Test)

# This connection will do for database-independent bug reports.
ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
ActiveRecord::Base.logger = Logger.new(STDOUT)

ActiveRecord::Schema.define do
  create_table :companies, force: true do |t|
    t.string :type
  end
end

class Company < ActiveRecord::Base; end
class Firm < Company; end
class Client < Company; end
class PriorityClient < Client; end

class BugTest < Minitest::Test
  def test_association_stuff
    puts Company.all.to_sql
    puts Firm.all.to_sql
    puts Client.all.to_sql
    puts PriorityClient.all.to_sql
  end
end
```

Note all of the subclassed models use `WHERE IN` like the STI we do here. Have you tried calling subclasses directly with your change? Does it even work? (`ManageIQ::Providers::Amazon::CloudManager::Vm.all`)