@kbrock 2017-12-06
>Have you tried calling subclasses directly with your change? Does it even work? 

Yes, I used vmware's vm class. See `Vmware::Vm` above - I abbreviated `ManageIQ::Providers::Vmware::InfraManager::Vm`. And to your point, abbreviations can be confusing.

While not an STI fan, I think it works. But for the `Vm` and `MiqTemplate` case using STI is a little suspect. I'd argue that they should be a top level construct. (Yes, being able to migrate across the types makes this not practical) Tacking on an extra 762 characters, and introducing an IN clause is a little gross. You especially feel lit when trying to debug an `Rbac` query.