@kbrock 2017-12-10
Current issue:

We always want `use_type_condition?` to return true. This gets tricky with the parent classes.

Returning `type_condition => nil` works well. but raises an error
`type_condition => arel.build_true` adds a bunch of true statements to the code